+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All Notable changes to `Csv` will be documented in this file
 
+## [Next] - TBD
+
+### Added
+
+- `SwapDelimiter` stream filter to allow working with multibyte CSV delimiter
+- `League\Csv\Serializer\AfterMapping` to work around the limitation aroud constructor usage.
+
+### Deprecated
+
+- None
+
+### Fixed
+
+- None
+
+### Removed
+
+- None
+
 ## [9.12.0](https://github.com/thephpleague/csv/compare/9.11.0...9.12.0) - 2023-12-02
 
 ### Added
@@ -1380,7 +1399,7 @@ to manage BOM character with CSV.
 
 Initial Release of `Bakame\Csv`
 
-[Next]: https://github.com/thephpleague/csv/compare/9.11.0...master
+[Next]: https://github.com/thephpleague/csv/compare/9.12.0...master
 [9.10.0]: https://github.com/thephpleague/csv/compare/9.9.0...9.10.0
 [9.9.0]: https://github.com/thephpleague/csv/compare/9.8.0...9.9.0
 [9.8.0]: https://github.com/thephpleague/csv/compare/9.7.4...9.8.0